+++ crates/burn-cube/src/lib.rs
@@ -3,24 +3,18 @@ extern crate alloc;
 #[macro_use]
 extern crate derive_new;
 
-// For use with *
-pub mod branch;
 pub mod codegen;
 
 mod compute;
-mod context;
-mod element;
-mod operation;
+mod language;
 mod pod;
 mod runtime;
 
 pub use codegen::*;
 pub use compute::*;
-pub use context::*;
-pub use element::*;
-pub use operation::*;
 pub use pod::*;
 pub use runtime::*;
+pub use language::*;
 
 pub use burn_cube_macros::cube;
