+++ crates/burn-cube/tests/language/function_call.rs
