+++ crates/burn-cube/tests/language/generic_kernel.rs
