+++ crates/burn-cube/tests/language/mod.rs
@@ -1,0 +1,13 @@
+mod cast_elem;
+mod cast_kind;
+mod for_loop;
+mod function_call;
+mod generic_kernel;
+mod r#if;
+mod if_else;
+mod literal;
+mod r#loop;
+mod module_import;
+mod parenthesis;
+mod reuse;
+mod r#trait;
